Moved event card component to home page (as events page may look different
