+++ frontend-site/src/components/Home/EventCard.vue
